create a type file for input file (#2)
