+++ components/auth/input-field/PhotoUpload.tsx
