+++ types/auth.ts
@@ -1,0 +1,9 @@
+export interface IFormData {
+  name: string;
+  email: string;
+  password: string;
+  confirmPassword: string;
+  birthday: string;
+  photo: File | null;
+}
+
